requirements.txt
streamlit==1.32.0
openai>=1.0.0
Pillow==10.2.0
python-dotenv==1.0.1
instabot==0.117.0
pandas==2.2.1
httpx>=0.24.0
pydantic>=2.0.0 